Remove use of makeStructuralReturn. NFC
The only caller of this function was __cxa_find_matching_catch.  IIRC
this function was designed for returning 64bit values by splitting them
into two 32bit values and returning one in tempRet0.  Historically it
was used like this:

 {{{ makeStructuralReturn(['low', 'hish']) }}};

See #12838 for an example of this kind of usage that no longer exists
in the codebase.

However, the final remaining caller (`__cxa_find_matching_catch`) is not
using it in this way.  `__cxa_find_matching_catch` return a single
pointer type and is expected to set tempRet0 (to an unrelated value).

This change makes this more explicit and removes the need to this
(misnamed IMHO) helper function.